Update sorts.rb
Insertion sort

Insertion sort removes one element from the input data, finds the location it belongs within the sorted list and inserts it there. It repeats until no input elements remain. In this code, first we set a variable equal to zero for it to be used as a placeholder for the current element in the iteration as we compare all of the elements in the array. We use i as the initial iteration variable. Because arrays are zero indexed, we need to take the length of the array and subtract 1 to determine the index number of the last spot in the array. Next we set two variables, j becomes the current value of i and we set a boolean in the variable done to false. We will then iterate through the array using a while loop and once the array has been sorted, done will be set to true and the loop will end. So, while j (the number in the current iteration) is greater than zero and the done variable is set to false, we first look at the number compare variable and add 1 to it. If the number in the jth position of the array is less than the number in the jth position of the array -1, we decrement the number currently saved in j by 1 and use that as the index number for the current position array. Then we take the j -1 position of the array and set it to the jth position of the array, and then save the value in the temporary variable and assign it to arr[j]. Once this has taken place with all of the numbers in the array, the done variable is set to true,  j is then set to j -1 and the loop completes and the sorted array is returned.

Selection sort

Selection sort compares the entire array for the smallest item and moves it to the first position in the list. Then it searches for the smallest item from the second position to the end of the list and moves that value to the second position and so on and so forth until the entire array is sorted. The values are switched. For each element i in the array range from 0 to the length of the array - 2, we set that value to the minimum_index variable. Inside of that loop we take each element j from the range of i + 1 to the length of the array -1. If the value stored in j is less than the value stored in minimum_index, the value in j becomes the new value of minimum_index, then we swap the value in i for the value in minimum_index.